refactor: SampleController를 HTTP Status 목적별로 분리 (#32)
- SampleController: 정상 비즈니스 로직 (200, 404, 409)
- SampleErrorController: 입력/인증/비즈니스 예외 (400, 401, 403, 422, 500)
- SampleGatewayController: Gateway/처리상태 예외 (202, 502, 503, 504)
- 테스트 클래스도 컨트롤러별로 분리 (일반/Debug 각 3개)
