+++ src/main/java/com/example/exception/playground/sample/adapter/in/web/SampleController.java
@@ -1,14 +1,7 @@
 package com.example.exception.playground.sample.adapter.in.web;
 
-import com.example.exception.playground.global.exception.AccessDeniedException;
-import com.example.exception.playground.global.exception.BusinessRuleViolationException;
 import com.example.exception.playground.global.exception.DuplicateResourceException;
-import com.example.exception.playground.global.exception.GatewayErrorException;
-import com.example.exception.playground.global.exception.GatewayTimeoutException;
 import com.example.exception.playground.global.exception.NotFoundException;
-import com.example.exception.playground.global.exception.RequestInProgressException;
-import com.example.exception.playground.global.exception.ServiceUnavailableException;
-import com.example.exception.playground.global.exception.UnauthorizedException;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -33,63 +26,5 @@ public class SampleController {
             throw new DuplicateResourceException("Sample with name 'duplicate' already exists");
         }
         return ResponseEntity.ok(Map.of("result", "created", "name", request.name()));
-    }
-
-    @GetMapping("/type-mismatch")
-    public ResponseEntity<Map<String, Object>> typeMismatch(@RequestParam Integer number) {
-        return ResponseEntity.ok(Map.of("number", number));
-    }
-
-    @GetMapping("/missing-param")
-    public ResponseEntity<Map<String, String>> missingParam(@RequestParam String required) {
-        return ResponseEntity.ok(Map.of("required", required));
-    }
-
-    @GetMapping("/unauthorized")
-    public ResponseEntity<Void> unauthorized() {
-        throw new UnauthorizedException("Invalid or expired token");
-    }
-
-    @GetMapping("/access-denied")
-    public ResponseEntity<Void> accessDenied() {
-        throw new AccessDeniedException("Insufficient permissions to access this resource");
-    }
-
-    @PostMapping("/business-rule")
-    public ResponseEntity<Map<String, String>> businessRule(@Valid @RequestBody SampleRequest request) {
-        if (request.age() > 100) {
-            throw new BusinessRuleViolationException("Age cannot exceed 100 for this operation");
-        }
-        return ResponseEntity.ok(Map.of("result", "ok"));
-    }
-
-    @GetMapping("/unexpected-error")
-    public ResponseEntity<Void> unexpectedError() {
-        throw new RuntimeException("Something went terribly wrong");
-    }
-
-    @GetMapping("/gateway-error")
-    public ResponseEntity<Void> gatewayError() {
-        throw new GatewayErrorException("Payment service connection refused");
-    }
-
-    @GetMapping("/gateway-timeout")
-    public ResponseEntity<Void> gatewayTimeout() {
-        throw new GatewayTimeoutException("Payment service did not respond within 5000ms");
-    }
-
-    @GetMapping("/service-unavailable")
-    public ResponseEntity<Void> serviceUnavailable() {
-        throw new ServiceUnavailableException("Payment service is under maintenance", 30);
-    }
-
-    @GetMapping("/service-unavailable-no-retry")
-    public ResponseEntity<Void> serviceUnavailableNoRetry() {
-        throw new ServiceUnavailableException("Payment service is temporarily unavailable");
-    }
-
-    @GetMapping("/request-in-progress")
-    public ResponseEntity<Void> requestInProgress() {
-        throw new RequestInProgressException("Request is being processed by payment service");
     }
 }
