+++ src/test/java/com/example/exception/playground/global/handler/GlobalExceptionHandlerDebugTest.java
@@ -17,7 +17,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @WebMvcTest(SampleController.class)
 @TestPropertySource(properties = "app.debug=true")
-@DisplayName("Debug Mode - ErrorResponse includes debugMessage and stackTrace")
+@DisplayName("Debug Mode - SampleController")
 class GlobalExceptionHandlerDebugTest {
 
     @Autowired
@@ -56,79 +56,6 @@ class GlobalExceptionHandlerDebugTest {
                     .andExpect(jsonPath("$.code").value("C006"))
                     .andExpect(jsonPath("$.traceId").exists())
                     .andExpect(jsonPath("$.debugMessage").value("NotFoundException: Sample with id 0 not found"))
-                    .andExpect(jsonPath("$.stackTrace").exists());
-        }
-    }
-
-    @Nested
-    @DisplayName("Unexpected Error in debug mode")
-    class UnexpectedErrorDebug {
-
-        @Test
-        @DisplayName("RuntimeException includes full debug info")
-        void unexpectedErrorWithDebug() throws Exception {
-            mockMvc.perform(get("/api/samples/unexpected-error"))
-                    .andDo(print())
-                    .andExpect(status().isInternalServerError())
-                    .andExpect(jsonPath("$.code").value("C999"))
-                    .andExpect(jsonPath("$.traceId").exists())
-                    .andExpect(jsonPath("$.debugMessage").value("RuntimeException: Something went terribly wrong"))
-                    .andExpect(jsonPath("$.stackTrace").exists());
-        }
-    }
-
-    @Nested
-    @DisplayName("Gateway Error in debug mode")
-    class GatewayErrorDebug {
-
-        @Test
-        @DisplayName("GatewayErrorException includes debug info with retryable fields")
-        void gatewayErrorWithDebug() throws Exception {
-            mockMvc.perform(get("/api/samples/gateway-error"))
-                    .andDo(print())
-                    .andExpect(status().isBadGateway())
-                    .andExpect(jsonPath("$.code").value("G001"))
-                    .andExpect(jsonPath("$.retryable").value(true))
-                    .andExpect(jsonPath("$.retryStrategy").value("RESUBMIT"))
-                    .andExpect(jsonPath("$.debugMessage").value("GatewayErrorException: Payment service connection refused"))
-                    .andExpect(jsonPath("$.stackTrace").exists());
-        }
-
-        @Test
-        @DisplayName("GatewayTimeoutException includes debug info")
-        void gatewayTimeoutWithDebug() throws Exception {
-            mockMvc.perform(get("/api/samples/gateway-timeout"))
-                    .andDo(print())
-                    .andExpect(status().isGatewayTimeout())
-                    .andExpect(jsonPath("$.code").value("G002"))
-                    .andExpect(jsonPath("$.retryable").value(true))
-                    .andExpect(jsonPath("$.debugMessage").value("GatewayTimeoutException: Payment service did not respond within 5000ms"))
-                    .andExpect(jsonPath("$.stackTrace").exists());
-        }
-
-        @Test
-        @DisplayName("ServiceUnavailableException includes debug info")
-        void serviceUnavailableWithDebug() throws Exception {
-            mockMvc.perform(get("/api/samples/service-unavailable"))
-                    .andDo(print())
-                    .andExpect(status().isServiceUnavailable())
-                    .andExpect(jsonPath("$.code").value("G003"))
-                    .andExpect(jsonPath("$.retryable").value(true))
-                    .andExpect(jsonPath("$.debugMessage").value("ServiceUnavailableException: Payment service is under maintenance"))
-                    .andExpect(jsonPath("$.stackTrace").exists())
-                    .andExpect(header().string("Retry-After", "30"));
-        }
-
-        @Test
-        @DisplayName("RequestInProgressException includes debug info with POLL_STATUS")
-        void requestInProgressWithDebug() throws Exception {
-            mockMvc.perform(get("/api/samples/request-in-progress"))
-                    .andDo(print())
-                    .andExpect(status().isAccepted())
-                    .andExpect(jsonPath("$.code").value("P001"))
-                    .andExpect(jsonPath("$.retryable").value(true))
-                    .andExpect(jsonPath("$.retryStrategy").value("POLL_STATUS"))
-                    .andExpect(jsonPath("$.debugMessage").value("RequestInProgressException: Request is being processed by payment service"))
                     .andExpect(jsonPath("$.stackTrace").exists());
         }
     }
